refactor：修改role、routes接口文件所在的目录
